+++ backend/src/services/downloaders/MissAVDownloader.ts
@@ -16,17 +16,12 @@ import {
 import { logger } from "../../utils/logger";
 import { ProgressTracker } from "../../utils/progressTracker";
 import {
-  ensureDirSafeSync,
   pathExistsSafeSync,
-  pathExistsTrustedSync,
   resolveSafeChildPath,
   statSafeSync,
   writeFileSafeSync,
 } from "../../utils/security";
-import { applyDedupeToRelatedPaths, dedupeRelativePath } from "../filenameTemplate/dedupe";
-import { planVideoOutputPaths } from "../filenameTemplate/renderer";
-import { enrichSourceOptionsForDownload } from "../filenameTemplate/sourceOptions";
-import { FilenameTemplateContext, FilenameTemplateSourceOptions } from "../filenameTemplate/types";
+import { FilenameTemplateSourceOptions } from "../filenameTemplate/types";
 import {
   flagsToArgs,
   getAxiosProxyConfig,
@@ -39,354 +34,18 @@ import { syncMediaServerArtifactsForRecord } from "../mediaServerExport";
 import * as storageService from "../storageService";
 import { Video } from "../storageService";
 import { BaseDownloader, DownloadOptions, VideoInfo } from "./BaseDownloader";
-
-const YT_DLP_PATH = process.env.YT_DLP_PATH || "yt-dlp";
-const MISSAV_PROGRESS_LOG_INTERVAL_MS = 10_000;
-const ALLOWED_MISSAV_LANGUAGE_SEGMENTS = new Set(["en", "ja", "zh", "ko"]);
-const ALLOWED_ROUTED_VIDEO_LANGUAGE_SEGMENTS = new Set([
-  ...ALLOWED_MISSAV_LANGUAGE_SEGMENTS,
-  "th",
-  "ms",
-  "de",
-  "fr",
-  "vi",
-  "id",
-  "fil",
-  "hi",
-]);
-const MISSAV_NAVIGATION_ORIGINS: Record<string, string> = {
-  "missav.com": "https://missav.com",
-  "missav.ai": "https://missav.ai",
-  "missav.ws": "https://missav.ws",
-  "missav.live": "https://missav.live",
-  "123av.com": "https://123av.com",
-  "123av.ai": "https://123av.ai",
-  "123av.ws": "https://123av.ws",
-  "javxx.com": "https://javxx.com",
-  "njavtv.com": "https://njavtv.com",
-};
-const PUPPETEER_MACOS_EXECUTABLE_PATHS = [
-  "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome",
-  "/Applications/Chromium.app/Contents/MacOS/Chromium",
-];
-const PUPPETEER_LINUX_EXECUTABLE_PATHS = [
-  "/usr/bin/google-chrome-stable",
-  "/usr/bin/google-chrome",
-  "/usr/bin/chromium-browser",
-  "/usr/bin/chromium",
-];
-const MISSAV_BROWSER_USER_AGENT =
-  "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/136.0.0.0 Safari/537.36";
-const MISSAV_BROWSER_ACCEPT_LANGUAGE = "en-US,en;q=0.9";
-const MISSAV_ROUTE_PREFIX_PATTERN = /^[a-zA-Z0-9_-]{2,20}$/;
-const MISSAV_CLOUDFLARE_CHALLENGE_PATTERN =
-  /cf-turnstile|Just a moment|security verification|challenge-platform/i;
-
-function resolvePuppeteerExecutablePath(): string | undefined {
-  const overridePath = process.env.PUPPETEER_EXECUTABLE_PATH?.trim();
-  if (overridePath) {
-    return overridePath;
-  }
-
-  const windowsPaths = [
-    process.env["PROGRAMFILES"]
-      ? path.join(
-          process.env["PROGRAMFILES"],
-          "Google",
-          "Chrome",
-          "Application",
-          "chrome.exe",
-        )
-      : null,
-    process.env["PROGRAMFILES(X86)"]
-      ? path.join(
-          process.env["PROGRAMFILES(X86)"],
-          "Google",
-          "Chrome",
-          "Application",
-          "chrome.exe",
-        )
-      : null,
-    process.env.LOCALAPPDATA
-      ? path.join(
-          process.env.LOCALAPPDATA,
-          "Google",
-          "Chrome",
-          "Application",
-          "chrome.exe",
-        )
-      : null,
-  ].filter((candidate): candidate is string => Boolean(candidate));
-
-  const candidatePaths =
-    process.platform === "darwin"
-      ? PUPPETEER_MACOS_EXECUTABLE_PATHS
-      : process.platform === "win32"
-        ? windowsPaths
-        : PUPPETEER_LINUX_EXECUTABLE_PATHS;
-
-  const resolvedPath = candidatePaths.find((candidatePath) =>
-    pathExistsTrustedSync(candidatePath),
-  );
-  if (resolvedPath) {
-    logger.info(`Using system Chrome for Puppeteer: ${resolvedPath}`);
-  }
-
-  return resolvedPath;
-}
-
-function resolvePuppeteerHeadlessMode(): boolean {
-  const override = process.env.PUPPETEER_HEADLESS?.trim().toLowerCase();
-  if (override === "false" || override === "0" || override === "no") {
-    return false;
-  }
-
-  return true;
-}
-
-function getMissAvPuppeteerLaunchOptions(): Parameters<typeof puppeteer.launch>[0] {
-  return {
-    headless: resolvePuppeteerHeadlessMode(),
-    executablePath: resolvePuppeteerExecutablePath(),
-    defaultViewport: {
-      width: 1280,
-      height: 900,
-    },
-    args: [
-      "--no-sandbox",
-      "--disable-setuid-sandbox",
-      "--disable-blink-features=AutomationControlled",
-      "--window-size=1280,900",
-      `--user-agent=${MISSAV_BROWSER_USER_AGENT}`,
-    ],
-  };
-}
-
-async function configureMissAvPage(page: {
-  setExtraHTTPHeaders?: (headers: Record<string, string>) => Promise<unknown>;
-  evaluateOnNewDocument?: (fn: () => void) => Promise<unknown>;
-}): Promise<void> {
-  await page.setExtraHTTPHeaders?.({
-    "accept-language": MISSAV_BROWSER_ACCEPT_LANGUAGE,
-  });
-  await page.evaluateOnNewDocument?.(() => {
-    Object.defineProperty(navigator, "webdriver", {
-      get: () => false,
-      configurable: true,
-    });
-  });
-}
-
-async function navigateMissAvPage(
-  page: {
-    goto: (
-      url: string,
-      options: { waitUntil: "domcontentloaded"; timeout: number },
-    ) => Promise<unknown>;
-    title?: () => Promise<string>;
-    content?: () => Promise<string>;
-    waitForFunction?: (
-      pageFunction: () => boolean,
-      options: { timeout: number },
-    ) => Promise<unknown>;
-  },
-  safeNavigationUrl: string,
-): Promise<void> {
-  logger.info("Navigating to:", safeNavigationUrl);
-  await page.goto(safeNavigationUrl, {
-    waitUntil: "domcontentloaded",
-    timeout: 60000,
-  });
-
-  const title = typeof page.title === "function" ? await page.title() : "";
-  if (title === "Just a moment..." && typeof page.waitForFunction === "function") {
-    logger.info(
-      "Cloudflare verification page detected; waiting up to 30 s for automatic completion...",
-    );
-    try {
-      await page.waitForFunction(
-        () =>
-          document.title !== "Just a moment..." &&
-          !document.body.innerText.includes("Performing security verification"),
-        { timeout: 30000 },
-      );
-    } catch (error) {
-      const html = typeof page.content === "function" ? await page.content() : "";
-      if (isCloudflareChallengeHtml(html)) {
-        throw new Error(
-          "MissAV access is blocked by Cloudflare verification. Retry with PUPPETEER_HEADLESS=false if needed.",
-        );
-      }
-      throw error;
-    }
-  }
-}
-
-function isCloudflareChallengeHtml(html: string): boolean {
-  return MISSAV_CLOUDFLARE_CHALLENGE_PATTERN.test(html);
-}
-
-function stripTrailingExtension(value: string, extension: string): string {
-  return value.endsWith(extension) ? value.slice(0, -extension.length) : value;
-}
-
-function getCanonicalMissAvHost(hostname: string): string | null {
-  const normalized = hostname.toLowerCase();
-
-  if (normalized === "missav.com" || normalized.endsWith(".missav.com")) {
-    return "missav.com";
-  }
-  if (normalized === "missav.ai" || normalized.endsWith(".missav.ai")) {
-    return "missav.ai";
-  }
-  if (normalized === "missav.ws" || normalized.endsWith(".missav.ws")) {
-    return "missav.ws";
-  }
-  if (normalized === "missav.live" || normalized.endsWith(".missav.live")) {
-    return "missav.live";
-  }
-  if (normalized === "123av.com" || normalized.endsWith(".123av.com")) {
-    return "123av.com";
-  }
-  if (normalized === "123av.ai" || normalized.endsWith(".123av.ai")) {
-    return "123av.ai";
-  }
-  if (normalized === "123av.ws" || normalized.endsWith(".123av.ws")) {
-    return "123av.ws";
-  }
-  if (normalized === "javxx.com" || normalized.endsWith(".javxx.com")) {
-    return "javxx.com";
-  }
-  if (normalized === "njavtv.com" || normalized.endsWith(".njavtv.com")) {
-    return "njavtv.com";
-  }
-
-  return null;
-}
-
-function usesRoutedVideoPath(canonicalHost: string): boolean {
-  return canonicalHost.startsWith("123av.") || canonicalHost === "javxx.com";
-}
-
-function buildSafeMissAvNavigationTarget(url: string): {
-  origin: string;
-  path: string;
-  url: string;
-} {
-  const parsedUrl = new URL(url);
-  if (parsedUrl.protocol !== "http:" && parsedUrl.protocol !== "https:") {
-    throw new Error(`Unsupported protocol for MissAV URL: ${parsedUrl.protocol}`);
-  }
-  if (parsedUrl.username || parsedUrl.password || parsedUrl.port) {
-    throw new Error(
-      "SSRF protection: URLs with credentials or explicit ports are not allowed.",
-    );
-  }
-
-  const canonicalHost = getCanonicalMissAvHost(parsedUrl.hostname);
-  if (!canonicalHost) {
-    throw new Error(`SSRF protection: Hostname ${parsedUrl.hostname} is not allowed.`);
-  }
-
-  const pathSegments = parsedUrl.pathname.split("/").filter(Boolean);
-  if (pathSegments.some((segment) => segment === "..")) {
-    throw new Error("SSRF protection: Path traversal is not allowed in URL path.");
-  }
-
-  const videoId = pathSegments[pathSegments.length - 1];
-  if (!videoId || !/^[a-zA-Z0-9_-]{2,120}$/.test(videoId)) {
-    throw new Error(
-      `SSRF protection: Invalid MissAV video path in URL: ${parsedUrl.pathname}`,
-    );
-  }
-
-  if (
-    usesRoutedVideoPath(canonicalHost) &&
-    pathSegments[pathSegments.length - 2]?.toLowerCase() === "v"
-  ) {
-    const prefixSegments = pathSegments.slice(0, -2);
-    if (prefixSegments.length > 1) {
-      throw new Error(
-        `SSRF protection: Invalid routed video path in URL: ${parsedUrl.pathname}`,
-      );
-    }
-
-    const normalizedRouteLanguage =
-      prefixSegments.length === 1 ? prefixSegments[0].toLowerCase() : null;
-    if (
-      normalizedRouteLanguage &&
-      !ALLOWED_ROUTED_VIDEO_LANGUAGE_SEGMENTS.has(normalizedRouteLanguage)
-    ) {
-      throw new Error(
-        `SSRF protection: Invalid routed video language segment in URL: ${parsedUrl.pathname}`,
-      );
-    }
-
-    const encodedVideoId = encodeURIComponent(videoId);
-    const safeOrigin = MISSAV_NAVIGATION_ORIGINS[canonicalHost];
-    if (!safeOrigin) {
-      throw new Error(
-        `SSRF protection: Hostname ${canonicalHost} has no allowed navigation origin.`,
-      );
-    }
-
-    const path = normalizedRouteLanguage
-      ? `/${normalizedRouteLanguage}/v/${encodedVideoId}`
-      : `/v/${encodedVideoId}`;
-    return {
-      origin: safeOrigin,
-      path,
-      url: `${safeOrigin}${path}`,
-    };
-  }
-
-  const maybeLanguage = pathSegments[pathSegments.length - 2]?.toLowerCase();
-  const normalizedLanguage =
-    maybeLanguage && ALLOWED_MISSAV_LANGUAGE_SEGMENTS.has(maybeLanguage)
-      ? maybeLanguage
-      : null;
-  const prefixSegments = normalizedLanguage
-    ? pathSegments.slice(0, -2)
-    : pathSegments.slice(0, -1);
-  if (prefixSegments.length > 1) {
-    throw new Error(
-      `SSRF protection: Invalid MissAV video path in URL: ${parsedUrl.pathname}`,
-    );
-  }
-
-  let normalizedRoutePrefix: string | null = null;
-  if (prefixSegments.length === 1) {
-    const candidatePrefix = prefixSegments[0]?.toLowerCase();
-    if (!candidatePrefix || !MISSAV_ROUTE_PREFIX_PATTERN.test(candidatePrefix)) {
-      throw new Error(
-        `SSRF protection: Invalid MissAV route prefix in URL: ${parsedUrl.pathname}`,
-      );
-    }
-    normalizedRoutePrefix = candidatePrefix;
-  }
-
-  const encodedVideoId = encodeURIComponent(videoId);
-  const safePath = normalizedRoutePrefix
-    ? normalizedLanguage
-      ? `/${normalizedRoutePrefix}/${normalizedLanguage}/${encodedVideoId}`
-      : `/${normalizedRoutePrefix}/${encodedVideoId}`
-    : normalizedLanguage
-      ? `/${normalizedLanguage}/${encodedVideoId}`
-      : `/${encodedVideoId}`;
-
-  const safeOrigin = MISSAV_NAVIGATION_ORIGINS[canonicalHost];
-  if (!safeOrigin) {
-    throw new Error(
-      `SSRF protection: Hostname ${canonicalHost} has no allowed navigation origin.`,
-    );
-  }
-
-  return {
-    origin: safeOrigin,
-    path: safePath,
-    url: `${safeOrigin}${safePath}`,
-  };
-}
+import { MISSAV_PROGRESS_LOG_INTERVAL_MS, YT_DLP_PATH } from "./missav/constants";
+import {
+  buildSafeMissAvNavigationTarget,
+  isCloudflareChallengeHtml,
+} from "./missav/navigation";
+import {
+  configureMissAvPage,
+  getMissAvPuppeteerLaunchOptions,
+  navigateMissAvPage,
+} from "./missav/puppeteer";
+import { selectBestM3u8Url } from "./missav/m3u8";
+import { planMissAvOutputPaths } from "./missav/outputPaths";
 
 export class MissAVDownloader extends BaseDownloader {
   // Implementation of IDownloader.getVideoInfo
@@ -660,117 +319,21 @@ export class MissAVDownloader extends BaseDownloader {
 
       // 6. Compute output paths using template or legacy formatter
       const settings = storageService.getSettings();
-      const moveThumbnailsToVideoFolder = settings.moveThumbnailsToVideoFolder || false;
-      const presetId = settings.downloadFilenamePresetId || "legacy";
-
-      let finalVideoFilename: string;
-      let finalThumbnailFilename: string;
-      let newVideoPath: string;
-      let newThumbnailPath: string;
-      let finalVideoWebPath: string;
-      let finalThumbnailWebPath: string | null;
-
-      if (presetId !== "legacy") {
-        // Non-legacy: use path planner
-        const uploadDateClean = videoDate.replace(/[^0-9]/g, "").slice(0, 8);
-        const year = uploadDateClean.length >= 4 ? uploadDateClean.slice(0, 4) : String(new Date().getFullYear());
-        const month = uploadDateClean.length >= 6 ? uploadDateClean.slice(4, 6) : String(new Date().getMonth() + 1).padStart(2, "0");
-        const day = uploadDateClean.length >= 8 ? uploadDateClean.slice(6, 8) : String(new Date().getDate()).padStart(2, "0");
-
-        const srcOpts = enrichSourceOptionsForDownload(
-          filenameTemplateSourceOptions || {},
-          {
-            author: videoAuthor,
-            uploadDate: videoDate,
-          }
-        );
-        const ctx: FilenameTemplateContext = {
-          title: videoTitle,
-          id: "",
-          ext: "",
-          uploader: videoAuthor,
-          channel: videoAuthor,
-          uploadDate: uploadDateClean,
-          uploadYear: year,
-          uploadMonth: month,
-          uploadDay: day,
-          durationSeconds: undefined,
-          durationString: "00-00",
-          artistName: videoAuthor,
-          sourceCustomName: srcOpts.sourceCustomName || videoAuthor,
-          sourceCollectionName: srcOpts.sourceCollectionName || videoAuthor,
-          sourceCollectionId: srcOpts.sourceCollectionId || "",
-          sourceCollectionType: srcOpts.sourceCollectionType || "single",
-          mediaPlaylistIndex: srcOpts.mediaPlaylistIndex,
-          mediaPlaylistIndexWithinDate: srcOpts.mediaPlaylistIndexWithinDate,
-          platform: "missav",
-          sourceUrl: url,
-        };
-
-        const planned = planVideoOutputPaths({
-          settings,
-          context: ctx,
-          videoExtension: mergeOutputFormat,
-          thumbnailExtension: "jpg",
-          moveThumbnailsToVideoFolder,
-          moveSubtitlesToVideoFolder: settings.moveSubtitlesToVideoFolder || false,
-        });
-
-        const reserved = new Set<string>();
-        const deduped = dedupeRelativePath(planned.video.relativePath, VIDEOS_DIR, reserved);
-        const { thumbnail: dedupedThumb } = applyDedupeToRelatedPaths(
-          planned.video.relativePath,
-          deduped,
-          planned.thumbnail.relativePath,
-          planned.subtitle.baseNameWithoutLanguageOrExt,
-        );
-
-        finalVideoFilename = path.basename(deduped);
-        newVideoPath = resolveSafeChildPath(VIDEOS_DIR, deduped);
-        finalThumbnailFilename = path.basename(dedupedThumb);
-        finalVideoWebPath = `/videos/${deduped}`;
-
-        const thumbnailDir = moveThumbnailsToVideoFolder ? VIDEOS_DIR : IMAGES_DIR;
-        newThumbnailPath = resolveSafeChildPath(thumbnailDir, dedupedThumb);
-        finalThumbnailWebPath = moveThumbnailsToVideoFolder
-          ? `/videos/${dedupedThumb}`
-          : `/images/${dedupedThumb}`;
-
-        ensureDirSafeSync(path.dirname(newVideoPath), VIDEOS_DIR);
-        ensureDirSafeSync(path.dirname(newThumbnailPath), [IMAGES_DIR, VIDEOS_DIR]);
-      } else {
-        // Legacy mode: use formatVideoFilename
-        const newSafeBaseFilename = formatVideoFilename(videoTitle, videoAuthor, videoDate);
-        const newVideoFilename = `${newSafeBaseFilename}.${mergeOutputFormat}`;
-        const newThumbnailFilename = `${newSafeBaseFilename}.jpg`;
-
-        finalVideoFilename = newVideoFilename;
-        finalThumbnailFilename = newThumbnailFilename;
-        newVideoPath = resolveSafeChildPath(VIDEOS_DIR, finalVideoFilename);
-
-        const thumbnailDir = moveThumbnailsToVideoFolder ? VIDEOS_DIR : IMAGES_DIR;
-
-        // If file already exists (e.g. redownload), deduplicate the filename
-        if (pathExistsSafeSync(newVideoPath, VIDEOS_DIR)) {
-          let counter = 1;
-          const ext = `.${mergeOutputFormat}`;
-          const basePath = stripTrailingExtension(newVideoPath, ext);
-          const baseName = newSafeBaseFilename;
-          while (pathExistsSafeSync(`${basePath}_${counter}${ext}`, VIDEOS_DIR)) {
-            counter++;
-          }
-          newVideoPath = `${basePath}_${counter}${ext}`;
-          finalVideoFilename = `${baseName}_${counter}${ext}`;
-          finalThumbnailFilename = `${baseName}_${counter}.jpg`;
-          logger.info(`File exists, using deduplicated filename: ${finalVideoFilename}`);
-        }
-
-        newThumbnailPath = resolveSafeChildPath(thumbnailDir, finalThumbnailFilename);
-        finalVideoWebPath = `/videos/${finalVideoFilename}`;
-        finalThumbnailWebPath = moveThumbnailsToVideoFolder
-          ? `/videos/${finalThumbnailFilename}`
-          : `/images/${finalThumbnailFilename}`;
-      }
+      const {
+        finalVideoFilename,
+        finalThumbnailFilename,
+        newVideoPath,
+        newThumbnailPath,
+        finalVideoWebPath,
+        finalThumbnailWebPath,
+      } = planMissAvOutputPaths(settings, {
+        videoTitle,
+        videoAuthor,
+        videoDate,
+        url,
+        mergeOutputFormat,
+        filenameTemplateSourceOptions,
+      });
 
       // 7. Download the video using yt-dlp with the m3u8 URL
       logger.info("Downloading video from m3u8 URL using yt-dlp:", m3u8Url);
@@ -1157,107 +720,11 @@ export class MissAVDownloader extends BaseDownloader {
     }
   }
 
-  // Helper to select best m3u8 URL
+  // Helper to select best m3u8 URL (delegates to ./missav/m3u8).
   static selectBestM3u8Url(
     urls: string[],
     hasFormatSort: boolean,
   ): string | null {
-    if (urls.length === 0) return null;
-
-    const getUrlParts = (input: string): { hostname: string; pathname: string } => {
-      try {
-        const parsedUrl = new URL(input);
-        return {
-          hostname: parsedUrl.hostname.toLowerCase(),
-          pathname: parsedUrl.pathname,
-        };
-      } catch {
-        return { hostname: "", pathname: "" };
-      }
-    };
-
-    const sortedUrls = [...urls].sort((a, b) => {
-      const aParts = getUrlParts(a);
-      const bParts = getUrlParts(b);
-
-      // 1. Priority: surrit.com
-      const aIsSurrit =
-        aParts.hostname === "surrit.com" ||
-        aParts.hostname.endsWith(".surrit.com");
-      const bIsSurrit =
-        bParts.hostname === "surrit.com" ||
-        bParts.hostname.endsWith(".surrit.com");
-      if (aIsSurrit && !bIsSurrit) return -1;
-      if (!aIsSurrit && bIsSurrit) return 1;
-
-      // 2. Priority: Master playlist (playlist.m3u8 specifically for surrit, or general master)
-      // We generally prefer master playlists because they contain all variants, allowing yt-dlp to pick the best.
-      // The previous logic penalized master playlists without explicit resolution, which caused issues.
-      const aIsMaster =
-        aParts.pathname.endsWith("/playlist.m3u8") ||
-        aParts.pathname.includes("/master/");
-      const bIsMaster =
-        bParts.pathname.endsWith("/playlist.m3u8") ||
-        bParts.pathname.includes("/master/");
-
-      // If we are strictly comparing surrit URLs (both are surrit), we prefer the master playlist
-      // because it's the "cleanest" source.
-      if (aIsSurrit && bIsSurrit) {
-        const aIsPlaylistM3u8 = aParts.pathname.includes("playlist.m3u8");
-        const bIsPlaylistM3u8 = bParts.pathname.includes("playlist.m3u8");
-        if (aIsPlaylistM3u8 && !bIsPlaylistM3u8) return -1;
-        if (!aIsPlaylistM3u8 && bIsPlaylistM3u8) return 1;
-      }
-
-      // If format sort is enabled, we almost always want the master playlist
-      if (hasFormatSort) {
-        if (aIsMaster && !bIsMaster) return -1;
-        if (!aIsMaster && bIsMaster) return 1;
-      } else {
-        // If NO format sort, previously we preferred specific resolution.
-        // BUT, given the bug report where a 240p stream was picked over a master,
-        // we should probably trust the master playlist more particularly if the alternative is low quality.
-        // However, if we have a high quality specific stream (e.g. 720p/1080p explicit), that might be fine.
-        // Let's refine: If one is surrit master, pick it. (Handled by step 1 & surrit sub-logic)
-        // If neither is surrit, and one is master...
-        // If both are master or both are not master, compare resolution.
-      }
-
-      // 3. Priority: Resolution (detected from URL)
-      const aQuality = a.match(/(\d+p)/)?.[1] || "0p";
-      const bQuality = b.match(/(\d+p)/)?.[1] || "0p";
-      const aQualityNum = parseInt(aQuality) || 0;
-      const bQualityNum = parseInt(bQuality) || 0;
-
-      // If we have a significant resolution difference, we might prefer the higher one
-      // UNLESS one is a master playlist and the other is a low res specific one.
-      // If one is master (0p detected) and other is 240p, 0p (master) should win if it's likely to contain better streams.
-
-      // Updated Strategy:
-      // If both have resolution, compare them.
-      if (aQualityNum > 0 && bQualityNum > 0) {
-        return bQualityNum - aQualityNum; // Higher quality first
-      }
-
-      // If one is master (assumed 0p from URL) and other is specific resolution:
-      // If we are prioritizing master playlists (e.g. because of surrit or format sort), master wins.
-      // If we are NOT specifically prioritizing master, we still might want to prefer it over very low res (e.g. < 480p).
-      if (aIsMaster && bQualityNum > 0 && bQualityNum < 480) return -1; // Master wins over < 480p
-      if (bIsMaster && aQualityNum > 0 && aQualityNum < 480) return 1; // Master wins over < 480p
-
-      // Fallback: Default to higher number (so 720p wins over 0p/master if we didn't catch it above)
-      // This preserves 'best attempt' for specific high quality URLs if they exist not on surrit.
-      if (aQualityNum !== bQualityNum) {
-        return bQualityNum - aQualityNum;
-      }
-
-      // Final tie-breaker: prefer master if all else equal
-      if (aIsMaster && !bIsMaster) return -1;
-      if (!aIsMaster && bIsMaster) return 1;
-
-      return 0;
-    });
-
-    return sortedUrls[0];
+    return selectBestM3u8Url(urls, hasFormatSort);
   }
 }
